Extracts integration tests
