+++ src/validator/fetch_checksums.rs
@@ -1,14 +1,14 @@
 // Copyright 2024 Dotanuki Labs
 // SPDX-License-Identifier: MIT
 
-use crate::validator::models::{OfficialWrapperChecksum, Result};
+use crate::validator::models::OfficialWrapperChecksum;
 use anyhow::anyhow;
 
 static HOST: &str = "https://cdn.statically.io/gh";
 static GITHUB_REPO: &str = "gradle/wrapper-validation-action";
 static CHECKSUMS_COLLECTION: &str = "main/src/wrapper-checksums.json";
 
-pub fn fetch() -> Result<Vec<OfficialWrapperChecksum>> {
+pub fn fetch() -> anyhow::Result<Vec<OfficialWrapperChecksum>> {
     let releases_url = format!("{}/{}/{}", HOST, GITHUB_REPO, CHECKSUMS_COLLECTION);
 
     let raw_response = ureq::get(&releases_url)
