+++ src/validator/find_wrappers.rs
@@ -1,12 +1,12 @@
 // Copyright 2024 Dotanuki Labs
 // SPDX-License-Identifier: MIT
 
-use crate::validator::models::{LocalGradleWrapper, Result};
+use crate::validator::models::LocalGradleWrapper;
 use anyhow::{anyhow, ensure};
 use std::path::Path;
 use walkdir::{DirEntry, WalkDir};
 
-pub fn locate(raw_path: &str) -> Result<Vec<LocalGradleWrapper>> {
+pub fn find(raw_path: &str) -> anyhow::Result<Vec<LocalGradleWrapper>> {
     let mut found_wrappers: Vec<LocalGradleWrapper> = WalkDir::new(raw_path)
         .into_iter()
         .filter_map(|entry| entry.ok())
@@ -20,7 +20,7 @@ pub fn locate(raw_path: &str) -> Result<Vec<LocalGradleWrapper>> {
     Ok(found_wrappers)
 }
 
-fn parse_wrapper_info(entry: DirEntry) -> Result<LocalGradleWrapper> {
+fn parse_wrapper_info(entry: DirEntry) -> anyhow::Result<LocalGradleWrapper> {
     let raw_path = entry.path().to_string_lossy();
     let complete_path = format!("{}/gradle/wrapper/gradle-wrapper.jar", &raw_path);
     let wrapper_jar_path = Path::new(&complete_path);
@@ -32,15 +32,15 @@ fn parse_wrapper_info(entry: DirEntry) -> Result<LocalGradleWrapper> {
 
 #[cfg(test)]
 mod tests {
-    use crate::validator::local_projects::locate;
+    use crate::validator::find_wrappers::find;
     use crate::validator::models::LocalGradleWrapper;
 
     #[test]
     fn should_locate_gradle_wrappers() {
         let project_dir = std::env::current_dir().unwrap();
-        let test_data_dir = format!("{}/test_data/valid", &project_dir.to_string_lossy());
+        let test_data_dir = format!("{}/test-data/valid", &project_dir.to_string_lossy());
 
-        let found_wrappers = locate(&test_data_dir).unwrap();
+        let found_wrappers = find(&test_data_dir).unwrap();
 
         let gradle7_wrapper_path = format!("{}/gradle7/gradle/wrapper/gradle-wrapper.jar", &test_data_dir);
         let gradle7_wrapper_sha256 = "575098db54a998ff1c6770b352c3b16766c09848bee7555dab09afc34e8cf590";
@@ -60,7 +60,7 @@ mod tests {
         let project_dir = std::env::current_dir().unwrap();
         let scripts_dir = format!("{}/scripts", &project_dir.to_string_lossy());
 
-        let found_wrappers = locate(&scripts_dir);
+        let found_wrappers = find(&scripts_dir);
 
         assert!(found_wrappers.is_err());
     }
