+++ test-data/invalid/tampered/gradle/wrapper/gradle-wrapper.jar
