+++ test-data/valid/gradle7/gradle/wrapper/gradle-wrapper.properties
