+++ test-data/valid/gradle8/gradle/wrapper/gradle-wrapper.properties
